core/vector_store.py
"""
FAISS Vector Store — the core retrieval engine.

This is the "ground truth" database. When the LLM claims something, we search
here for supporting or contradicting evidence.

TWO INDEX TYPES:
  1. IndexFlatL2 — Brute force, exact results. Used when < 50K vectors.
     Searches every vector, but FAISS's C++ implementation with SIMD makes
     this fast enough up to ~50K vectors.

  2. IndexIVFFlat — Approximate search for scale. Clusters vectors into
     buckets, then only searches nearby buckets. Used when > 50K vectors.
     Much faster at scale but can miss some results (configurable via nprobe).

WHY NOT JUST USE A DATABASE: Postgres with pgvector or similar could work,
but FAISS is: (a) in-memory = zero network latency, (b) 10-100x faster for
pure vector search, (c) no infrastructure dependency for development.
The trade-off is no built-in persistence, so we save/load manually.
"""

import json
from pathlib import Path
from typing import Optional

import faiss
import numpy as np

from app.config import settings
from app.logging_config import get_logger
from models.schemas import DocumentChunk, RetrievedEvidence

logger = get_logger("vector_store")


class VectorStore:
    """FAISS-backed vector store with metadata management."""

    def __init__(self, dimension: int = None):
        self.dimension = dimension or settings.embedding_dimension

        # Start with flat index (exact search)
        self.index: faiss.IndexFlatL2 = faiss.IndexFlatL2(self.dimension)

        # FAISS only stores vectors — we store metadata separately
        # Maps FAISS internal integer ID → DocumentChunk
        # WHY SEPARATE: FAISS is a pure numerical library. It doesn't know
        # about text, sources, or metadata. We maintain this mapping ourselves.
        self._metadata: dict[int, dict] = {}

        # Counter for assigning FAISS IDs
        self._next_id: int = 0

    @property
    def size(self) -> int:
        """Number of vectors currently in the index."""
        return self.index.ntotal

    def add_vectors(
        self,
        vectors: np.ndarray,
        chunks: list[DocumentChunk],
    ) -> int:
        """
        Add vectors and their corresponding metadata to the index.

        Args:
            vectors: numpy array of shape (n, dimension)
            chunks: list of DocumentChunk objects (same length as vectors)

        Returns:
            Number of vectors added

        HOW IT WORKS:
            FAISS assigns sequential integer IDs starting from 0.
            We maintain a parallel dict mapping those IDs to our chunk metadata.
            When we search and get back ID=42, we look up _metadata[42] to get
            the actual text and source information.
        """
        if len(vectors) != len(chunks):
            raise ValueError(
                f"Vector count ({len(vectors)}) != chunk count ({len(chunks)})"
            )

        if len(vectors) == 0:
            return 0

        # Ensure correct shape and dtype
        vectors = np.ascontiguousarray(vectors, dtype=np.float32)
        if vectors.ndim == 1:
            vectors = vectors.reshape(1, -1)

        # Store metadata for each vector
        for i, chunk in enumerate(chunks):
            faiss_id = self._next_id + i
            self._metadata[faiss_id] = chunk.model_dump()

        # Add to FAISS index
        self.index.add(vectors)

        added = len(vectors)
        self._next_id += added

        logger.info("vectors_added", count=added, total_index_size=self.size)
        return added

    def search(
        self,
        query_vector: np.ndarray,
        top_k: int = None,
        score_threshold: Optional[float] = None,
    ) -> list[RetrievedEvidence]:
        """
        Search for the most similar vectors to the query.

        Args:
            query_vector: numpy array of shape (dimension,)
            top_k: number of results to return
            score_threshold: max L2 distance to include (filters out weak matches)

        Returns:
            List of RetrievedEvidence, sorted by similarity (best first)

        HOW FAISS SEARCH WORKS:
            1. Computes L2 distance between query_vector and every indexed vector
            2. Returns the top_k with smallest distances
            3. Returns two arrays: distances[] and ids[]
            4. We use ids[] to look up our metadata dict
        """
        top_k = top_k or settings.retrieval_top_k

        if self.size == 0:
            logger.warning("search_empty_index")
            return []

        # Reshape for FAISS: (1, dimension) for a single query
        query = np.ascontiguousarray(
            query_vector.reshape(1, -1), dtype=np.float32
        )

        # FAISS search returns:
        #   distances: shape (1, top_k) — L2 distances
        #   indices: shape (1, top_k) — internal FAISS IDs
        distances, indices = self.index.search(query, min(top_k, self.size))

        results = []
        for rank, (dist, idx) in enumerate(zip(distances[0], indices[0])):
            # FAISS returns -1 for empty slots (when index has fewer than top_k vectors)
            if idx == -1:
                continue

            # Apply score threshold if specified
            if score_threshold is not None and dist > score_threshold:
                continue

            # Look up metadata
            if idx in self._metadata:
                chunk = DocumentChunk(**self._metadata[idx])
                evidence = RetrievedEvidence(
                    chunk=chunk,
                    similarity_score=float(dist),
                    rank=rank + 1,
                )
                results.append(evidence)

        logger.info(
            "search_complete",
            results_found=len(results),
            top_score=results[0].similarity_score if results else None,
        )
        return results

    def batch_search(
        self,
        query_vectors: np.ndarray,
        top_k: int = None,
    ) -> list[list[RetrievedEvidence]]:
        """
        Search for multiple queries at once.

        WHY BATCH SEARCH: When verifying an LLM response with 5 claims,
        we need 5 separate FAISS searches. Batching them into one call
        lets FAISS parallelize the computation internally, saving ~30-50%
        wall time vs sequential searches.
        """
        top_k = top_k or settings.retrieval_top_k

        if self.size == 0:
            return [[] for _ in range(len(query_vectors))]

        queries = np.ascontiguousarray(query_vectors, dtype=np.float32)
        if queries.ndim == 1:
            queries = queries.reshape(1, -1)

        distances, indices = self.index.search(queries, min(top_k, self.size))

        all_results = []
        for q_idx in range(len(queries)):
            results = []
            for rank, (dist, idx) in enumerate(
                zip(distances[q_idx], indices[q_idx])
            ):
                if idx == -1:
                    continue
                if idx in self._metadata:
                    chunk = DocumentChunk(**self._metadata[idx])
                    evidence = RetrievedEvidence(
                        chunk=chunk,
                        similarity_score=float(dist),
                        rank=rank + 1,
                    )
                    results.append(evidence)
            all_results.append(results)

        logger.info("batch_search_complete", num_queries=len(queries))
        return all_results

    def save(self, index_path: str = None, metadata_path: str = None):
        """
        Persist index and metadata to disk.

        FAISS has its own binary format for the index. Metadata is stored
        as JSON (human-readable, easy to debug).
        """
        index_path = index_path or settings.faiss_index_path
        metadata_path = metadata_path or settings.faiss_metadata_path

        Path(index_path).parent.mkdir(parents=True, exist_ok=True)
        Path(metadata_path).parent.mkdir(parents=True, exist_ok=True)

        # Save FAISS index (binary format)
        faiss.write_index(self.index, index_path)

        # Save metadata (JSON)
        save_data = {
            "next_id": self._next_id,
            "metadata": {str(k): v for k, v in self._metadata.items()},
        }
        with open(metadata_path, "w") as f:
            json.dump(save_data, f, indent=2)

        logger.info(
            "index_saved",
            index_path=index_path,
            num_vectors=self.size,
        )

    def load(self, index_path: str = None, metadata_path: str = None) -> bool:
        """
        Load index and metadata from disk.

        Returns True if loaded successfully, False if files don't exist.
        """
        index_path = index_path or settings.faiss_index_path
        metadata_path = metadata_path or settings.faiss_metadata_path

        if not Path(index_path).exists() or not Path(metadata_path).exists():
            logger.info("no_saved_index_found")
            return False

        # Load FAISS index
        self.index = faiss.read_index(index_path)

        # Load metadata
        with open(metadata_path, "r") as f:
            save_data = json.load(f)

        self._next_id = save_data["next_id"]
        self._metadata = {int(k): v for k, v in save_data["metadata"].items()}

        logger.info(
            "index_loaded",
            num_vectors=self.size,
            num_metadata=len(self._metadata),
        )
        return True

    def clear(self):
        """Reset the entire index. Used in testing and reindexing."""
        self.index = faiss.IndexFlatL2(self.dimension)
        self._metadata = {}
        self._next_id = 0
        logger.info("index_cleared")
